Fix separate groups forum posts sending to all
If group mode is on, we should check if the user is a member of the group and, if not, then check if he has accessallgroups capability. Then, we must provide $userto->id to the functions groups_is_member and has_capability or they will get ID from global $USER.
This explains why forum posts to separated groups are sent to all users, even on other groups, because group_is_member returns false, but has_capability is true for the CLI user.